+++ CHANGES.md
@@ -1,9 +1,11 @@
-## Upcoming: 0.1.0
+## In the next release...
+
+## 0.0.4
 
 * **Breaking Change** `client`, `server`, and `max-rps` now subcommands of `strest-grpc` executable.
 * **Breaking Change** All CLI flags take two dashes now.
 * Add support for Unix Domain sockets.
-* Support `--errorRate` for streaming requests.
+* [client] Support `--errorRate` for streaming requests.
 
 ## 0.0.3